Update generation for removed variable type group classes
